# Improve performance and reduce allocations of Cache