… src/Ninject.Test/Unit/Activation/ContextTests.cs