> Marked **WIP** as I still need to update tests for **Context**, and want to increase test coverage for **Cache**.

**Cache:**
* Add overloads of **Remember** and **TryGet** that take scope.
* Fixed thread safety.
* Improve overal performance and reduce allocations.

**Context:**
* Use new `Remember(IContext context, object scope, InstanceReference reference)` and `TryGet(IContext context, object scope)` overloads on **Cache**.
* Only try to get instance from cache when there's a scope

**Before:**

|                                         Method |       Mean |     Error |     StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|----------------------------------------------- |-----------:|----------:|-----------:|------------:|------------:|------------:|--------------------:|
|                         TryGet_NoCacheForScope |   127.5 ns |  1.966 ns |   1.839 ns |           - |           - |           - |                   - |
|         TryGet_OneEntryForBindingConfiguration |   344.6 ns |  6.823 ns |   8.380 ns |      0.0093 |           - |           - |                40 B |
|      TryGet_ZeroEntriesForBindingConfiguration |   320.8 ns |  6.211 ns |   6.646 ns |      0.0093 |           - |           - |                40 B |
| TryGet_MoreThanOneEntryForBindingConfiguration |   345.3 ns |  6.699 ns |   6.879 ns |      0.0093 |           - |           - |                40 B |
|                       Remember_NoCacheForScope | 2,359.4 ns | 46.578 ns |  69.716 ns |      0.3438 |           - |           - |              1456 B |
|                                       Remember | 1,541.7 ns | 30.168 ns |  38.152 ns |      0.0371 |      0.0098 |           - |               240 B |
|        Release_InstanceNotFound_SingleThreaded | 2,684.2 ns | 55.521 ns | 113.415 ns |      0.2695 |           - |           - |              1136 B |
|         Release_InstanceNotFound_MultiThreaded | 4,088.8 ns | 27.446 ns |  24.330 ns |      0.2700 |           - |           - |                   - |
|              RememberAndRelease_Singlethreaded | 3,653.0 ns | 27.788 ns |  24.633 ns |      0.3624 |           - |           - |              1528 B |
|               RememberAndRelease_Multithreaded |         NA |        NA |         NA |           - |           - |           - |                   - |

Note:
The last benchmark fails due to a concurrency issue which is fixed by this PR.

**After:**


|                                         Method |        Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|----------------------------------------------- |------------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
|                         TryGet_NoCacheForScope |    60.90 ns | 0.1086 ns | 0.0907 ns |           - |           - |           - |                   - |
|         TryGet_OneEntryForBindingConfiguration |   208.53 ns | 0.2269 ns | 0.2011 ns |           - |           - |           - |                   - |
|      TryGet_ZeroEntriesForBindingConfiguration |   200.74 ns | 0.1229 ns | 0.1026 ns |           - |           - |           - |                   - |
| TryGet_MoreThanOneEntryForBindingConfiguration |   212.99 ns | 1.1049 ns | 0.8626 ns |           - |           - |           - |                   - |
|                       Remember_NoCacheForScope | 2,018.77 ns | 40.187 ns | 92.336 ns |      0.2852 |      0.0039 |           - |              1200 B |
|                                       Remember | 1,549.72 ns | 33.164 ns | 32.572 ns |      0.0273 |      0.0078 |           - |               168 B |
|        Release_InstanceNotFound_SingleThreaded |   993.08 ns | 19.669 ns | 29.440 ns |      0.0605 |           - |           - |               256 B |
|         Release_InstanceNotFound_MultiThreaded |   789.38 ns | 15.619 ns | 16.040 ns |      0.0614 |           - |           - |                   - |
|              RememberAndRelease_Singlethreaded | 2,192.04 ns | 42.820 ns | 43.973 ns |      0.1450 |      0.0038 |           - |               616 B |
|               RememberAndRelease_Multithreaded | 3,003.57 ns | 56.312 ns | 57.829 ns |      0.1450 |      0.0050 |           - |                   - |